Fixed a MSVC6 warning in matrixlib. Added a "Public server" yes/no line in the server creation menu. Removed a duplicated cvar declaration in "net.h".
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@2821 d7cf8633-e32d-0410-b094-e92efae38249

+++ matrixlib.c
@@ -431,6 +431,6 @@ void Matrix4x4_OriginFromMatrix (const matrix4x4_t *in, float *out)
 float Matrix4x4_ScaleFromMatrix (const matrix4x4_t *in)
 {
 	// we only support uniform scaling, so assume the first row is enough
-	return sqrt(in->m[0][0] * in->m[0][0] + in->m[0][1] * in->m[0][1] + in->m[0][2] * in->m[0][2]);
+	return (float)sqrt(in->m[0][0] * in->m[0][0] + in->m[0][1] * in->m[0][1] + in->m[0][2] * in->m[0][2]);
 }
